Add files via upload
The goal of this backtest.py was to evaluate the performance of a Time-Weighted Average Price(TWAP) strategy for executing an order of 10,000 shares over a trading session. The TWAPstrategy divides the total order into 10 equal slices and executes each slice at regular intervals, aiming to minimize market impact and execution costs.